moved sparkml utils from where it caused depenency errors
