+++ onnxmltools/utils/__init__.py
@@ -17,4 +17,3 @@ from .tests_helper import dump_data_and_model
 from .tests_helper import dump_one_class_classification, dump_binary_classification, dump_multiple_classification
 from .tests_helper import dump_multiple_regression, dump_single_regression
 from .tests_dl_helper import create_tensor
-from .sparkml_test_utils import *
