+++ tests/sparkml/__init__.py
@@ -5,3 +5,4 @@
 # --------------------------------------------------------------------------
 
 from .sparkml_test_base import *
+from .sparkml_test_utils import start_spark, stop_spark, dump_data_and_sparkml_model,dataframe_to_nparray
