+++ tests/sparkml/test_logistic_regression.py
@@ -5,12 +5,10 @@ import unittest
 import numpy
 from pyspark.ml.classification import LogisticRegression
 from pyspark.ml.linalg import VectorUDT, SparseVector
-from pyspark.sql.types import ArrayType, FloatType
 
 from onnxmltools import convert_sparkml
 from onnxmltools.convert.common.data_types import FloatTensorType
-from onnxmltools.utils import dump_data_and_sparkml_model
-from sparkml import SparkMlTestCase
+from tests.sparkml import SparkMlTestCase, dump_data_and_sparkml_model
 
 
 class TestSparkmlLogisticRegression(SparkMlTestCase):
