+++ tests/sparkml/test_pipeline.py
@@ -9,8 +9,7 @@ from pyspark.ml.feature import StringIndexer, OneHotEncoderEstimator, VectorAsse
 
 from onnxmltools import convert_sparkml
 from onnxmltools.convert.common.data_types import StringTensorType
-from onnxmltools.utils import dump_data_and_sparkml_model
-from sparkml import SparkMlTestCase
+from tests.sparkml import SparkMlTestCase, dump_data_and_sparkml_model
 
 
 class TestSparkmlPipeline(SparkMlTestCase):
